+++ cmd/system.go
@@ -25,13 +25,14 @@ func init() {
 	deployOverrides = svc.DeployOverrides{}
 	deployConfig = cfg.DeployConfig{}
 	systemDeployCmd.Flags().StringVarP(&deployConfig.DeployLocalPath, "local", "l", "", "Path to a local DCE repo to deploy.")
-	systemDeployCmd.Flags().BoolVarP(&deployConfig.UseCached, "use-cached", "c", true, "Overwrite local backend state.")
+	systemDeployCmd.Flags().BoolVarP(&deployConfig.UseCached, "use-cached", "c", true, "Uses locally-cached files, if available.")
 	systemDeployCmd.Flags().BoolVarP(&deployConfig.BatchMode, "batch-mode", "b", false, "Skip prompting for resource creation.")
-	systemDeployCmd.Flags().StringVar(&deployConfig.TFInitOptions, "tf-init-options", "", "Options to pass to the underlying `tf init` command.")
-	systemDeployCmd.Flags().StringVar(&deployConfig.TFApplyOptions, "tf-apply-options", "", "Options to pass to the underlying `tf apply` command.")
+	systemDeployCmd.Flags().StringVar(&deployConfig.TFInitOptions, "tf-init-options", "", "Options to pass to the underlying \"tf init\" command.")
+	systemDeployCmd.Flags().StringVar(&deployConfig.TFApplyOptions, "tf-apply-options", "", "Options to pass to the underlying \"tf apply\" command.")
+	systemDeployCmd.Flags().BoolVar(&deployConfig.SaveTFOptions, "save-options", false, "If specified, saves the values provided by \"--tf-init-options\" and \"--tf-apply-options\" in the config file.")
 	systemDeployCmd.Flags().StringVarP(&deployOverrides.Namespace, "namespace", "n", "", "Set a custom terraform namespace (Optional)")
 	systemDeployCmd.Flags().StringVarP(&deployOverrides.AWSRegion, "region", "r", "", "The aws region that DCE will be deployed to (Default: us-east-1)")
-	systemDeployCmd.Flags().StringArrayVarP(&deployOverrides.GlobalTags, "tag", "t", []string{}, "Tags to be placed on all DCE resources. E.g. `dce system deploy --tag key1:value1 --tag key2:value2`")
+	systemDeployCmd.Flags().StringArrayVarP(&deployOverrides.GlobalTags, "tag", "t", []string{}, "Tags to be placed on all DCE resources. E.g. \"dce system deploy --tag key1:value1 --tag key2:value2\"")
 	systemDeployCmd.Flags().StringVar(&deployOverrides.BudgetNotificationFromEmail, "budget-notification-from-email", "", "Email address from which budget notifications will be sent")
 	systemDeployCmd.Flags().StringArrayVar(&deployOverrides.BudgetNotificationBCCEmails, "budget-notification-bcc-emails", []string{}, "Email address from which budget notifications will be sent")
 	systemDeployCmd.Flags().StringVar(&deployOverrides.BudgetNotificationTemplateHTML, "budget-notification-template-html", "", "HTML template for budget notification emails")
@@ -59,6 +60,7 @@ var systemDeployCmd = &cobra.Command{
 
 		ctx := context.WithValue(context.Background(), constants.DeployConfig, &deployConfig)
 		if err := Service.Deploy(ctx, &deployOverrides); err != nil {
+			cmd.SilenceUsage = true
 			return err
 		}
 		return nil