+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## vNext
 - Added `--tf-init-options` and `--tf-apply-options` for greater control over underlying provisioner
+- Added `--save-options` flag to persist the `--tf-init-options` and `--tf-apply-options` to the
+  configuration file when specified. Default behavior is false.
 - Persist API host name and path after running `system deploy`
 
 ## v0.3.1